chore: code cleanup to avoid too much magic in the bin/codeflare shell script
Push this complexity to a kui catchall command registration. We want `codeflare ml/codeflare` to open in guide mode. We do this, in this PR, by using a Kui catchall, catching all unknown command lines that start with `codeflare ...`. With this in place, we can remove some of the junk in the bin/codeflare shell script.

This PR also fixes a number of bugs with handling of spaces in filepaths. Why add this to this PR? Because we need to update bin/codeflare to use "$@" to fix spaces... This requires us to add `codeflare ` as a prefix, which leads to the catchall ...