feat: add Dashboard tray menu, with MLFlow entry
This menu item uses the ml/ray/start/kubernetes/port-forward/mlflow guidebook to set up a port forward to the mlflow service started by the codeflare ray helm chart. It opens a kui window to localhost:thatPort, and then tears down the port-forward when the user closes that popup window.