feat: bump to madwizard 1.0.0
BREAKING CHANGE: this version of madwizard updates the profile keys in a way that renders prior profiles invalid.

BREAKING CHANGE: this version of madwizared necessitates guidebook to use import rather than inlining for sub-guidebooks that offer the user a choice. this is done so that each choice can be easily scoped based on the filepath of that guidebook within the store hierarchy.