fix: ray helm chart should specify `parallelism` to avoid livelock
https://github.com/guidebooks/store/pull/634

if a Job does not specify parallelism=completions, then a livelock will occur. with the default parallelism (which is 1), the Job controller creates one Pod at a time, waiting till it is scheduled before creating the next one.
meanwhile, the coscheduler doesn’t allow that first one to be scheduled until the rest of the Pods are created…

and … for ray, we were using Jobs with default parallelism