fix: force yoga.wasm to be loaded as a file resource
it turns out that `yoga-wasm-web` does something silly with the way it loads the wasm; rather than a simple `import('./yoga.wasm')` it does some gymnastics that include a `readFile`, thus assuming the wasm is a file, not a more general module. so this change adds a webpack rule that forces just yoga.wasm to be treated as a file resource.